finished minimal tests
